tests: reorganize
